+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # pygradle news #
 
+* 2019-04-08
+  - You must now change any usages of either the `python.pex.FatPex` or
+    `python.pex.isFat` flag to `python.zipapp.isFat`.
+
 * 2018-11-09
   - Added a default `mypy` task which you can enable by setting
     `python.mypy.run = true` in your `build.gradle` file.